move transmit methods from wallet to lightclient and clean up lightclient send mod
