+++ zingolib/src/wallet.rs
@@ -134,15 +134,15 @@ pub struct LightWallet {
     pub shard_trees: ShardTrees,
     /// Sync state
     pub sync_state: SyncState,
-    /// Wallet settings.
+    /// Wallet settings
     pub wallet_settings: WalletSettings,
     /// The current and historical daily price of zec.
     pub price_list: PriceList,
     /// Progress of an outgoing transaction
     // TODO: move to LightClient
     pub send_progress: SendProgress,
-    /// Latest proposal
-    latest_proposal: Option<ZingoProposal>,
+    /// Send proposal
+    send_proposal: Option<ZingoProposal>,
     /// Boolean for tracking whether the wallet state has changed since last save.
     pub save_required: bool,
 }
@@ -257,7 +257,7 @@ impl LightWallet {
             wallet_settings,
             price_list: PriceList::new(),
             save_required: true,
-            latest_proposal: None,
+            send_proposal: None,
             send_progress: SendProgress::new(0),
         })
     }
@@ -337,8 +337,9 @@ impl LightWallet {
         )
     }
 
-    pub(crate) fn latest_proposal(&self) -> Option<&ZingoProposal> {
-        self.latest_proposal.as_ref()
+    /// Clears the proposal in the `send_proposal` field.
+    pub fn clear_proposal(&mut self) {
+        self.send_proposal = None;
     }
 
     #[must_use]
